Moving some code to a new utils.go file and start adding a shell exec with a timeout
