+++ driver_test.go
@@ -120,17 +120,6 @@ func TestSocketActivate(t *testing.T) {
 
 }
 
-func TestSh_success(t *testing.T) {
-	out, err := sh("ls")
-	assert.Nil(t, err, formatError("ls", err))
-	assert.Contains(t, out, "driver_test.go")
-}
-
-func TestSh_fail(t *testing.T) {
-	_, err := sh("false")
-	assert.NotNil(t, err, formatError("false", err))
-}
-
 // Helpers
 func formatError(name string, err error) string {
 	return fmt.Sprintf("ERROR calling %s: %q", name, err)
